requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic
python-dotenv
pytest
httpx
pytest-asyncio
black
isort
psycopg2-binary
PyJWT